improved versioning
