+++ src/sync/core/events/organization/deleted/organization-deleted-v1.event.ts
@@ -1,10 +1,10 @@
 import { Exclude } from 'class-transformer';
-import { OrganizationEvent } from '../../organization.event';
+import { OrganizationEvent } from '../organization.event';
 
 @Exclude()
 export class OrganizationDeleted extends OrganizationEvent {
 
-  static version = '1.0.0';
+  static version = '1';
 
   constructor(organizationId: string) {
     super(organizationId, OrganizationDeleted.version);
