+++ src/sync/core/events/organization/registered/organization-registered-v1.event.ts
@@ -1,10 +1,10 @@
 import { Exclude, Expose } from 'class-transformer';
-import { OrganizationEvent } from '../../organization.event';
+import { OrganizationEvent } from '../organization.event';
 
 @Exclude()
-export class OrganizationUpdated extends OrganizationEvent {
+export class OrganizationRegistered extends OrganizationEvent {
 
-  static version = '1.0.0';
+  static version = '1';
 
   @Expose()
   public readonly website: string;
@@ -20,7 +20,7 @@ export class OrganizationUpdated extends OrganizationEvent {
 
   constructor(organizationId: string, website: string, contactName: string,
               contactEmail: string, contactPhone: string) {
-    super(organizationId, OrganizationUpdated.version);
+    super(organizationId, OrganizationRegistered.version);
 
     this.website = website;
     this.contactName = contactName;
