+++ src/sync/core/events/sensor/activated/sensor-activated-v1.event.ts
@@ -1,10 +1,10 @@
 import { Exclude } from 'class-transformer';
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 
 @Exclude()
 export class SensorActivated extends SensorEvent {
 
-  static version = '1.0.0';
+  static version = '1';
 
   constructor(sensorId: string) {
     super(sensorId, SensorActivated.version);
