--- src/sync/core/events/sensor/ds-added/1.0.0/sensor-datastreamadded.event.ts
@@ -1,3 +1,0 @@
-import { DatastreamChanged } from '../../sensor-datastreamchanged.event';
-
-export class DatastreamAdded extends DatastreamChanged {}
