+++ src/sync/core/events/sensor/ds-added/sensor-datastreamadded-v1.event.ts
@@ -1,0 +1,3 @@
+import { DatastreamChanged } from '../sensor-datastreamchanged.event';
+
+export class DatastreamAdded extends DatastreamChanged {}
