+++ src/sync/core/events/sensor/ds-updated/index.ts
@@ -1,8 +1,8 @@
 import { plainToClass } from 'class-transformer';
 import { EventMessage } from '../../event-message';
-import { DatastreamUpdated as V1 } from './1.0.0/sensor-datastreamupdated.event';
+import { DatastreamUpdated as V1 } from './sensor-datastreamupdated-v1.event';
 
-export { DatastreamUpdated } from './1.0.0/sensor-datastreamupdated.event';
+export { DatastreamUpdated } from './sensor-datastreamupdated-v1.event';
 
 export function getDatastreamUpdatedEvent(eventMessage: EventMessage): V1 {
     return !eventMessage.metadata.version || eventMessage.metadata.version === V1.version ? plainToClass(V1, eventMessage.data) : null;
