+++ src/sync/core/events/sensor/ds-updated/sensor-datastreamupdated-v1.event.ts
@@ -1,0 +1,3 @@
+import { DatastreamChanged } from '../sensor-datastreamchanged.event';
+
+export class DatastreamUpdated extends DatastreamChanged {}
