+++ src/sync/core/events/sensor/ownership-transferred/index.ts
@@ -1,8 +1,8 @@
 import { plainToClass } from 'class-transformer';
 import { EventMessage } from '../../event-message';
-import { SensorOwnershipTransferred as V1 } from './1.0.0/sensor-ownershiptransferred.event';
+import { SensorOwnershipTransferred as V1 } from './sensor-ownershiptransferred-v1.event';
 
-export { SensorOwnershipTransferred } from './1.0.0/sensor-ownershiptransferred.event';
+export { SensorOwnershipTransferred } from './sensor-ownershiptransferred-v1.event';
 
 export function getSensorOwnershipTransferredEvent(eventMessage: EventMessage): V1 {
     return !eventMessage.metadata.version || eventMessage.metadata.version === V1.version ? plainToClass(V1, eventMessage.data) : null;
