+++ src/sync/core/events/sensor/registered/index.ts
@@ -1,9 +1,9 @@
 import { plainToClass } from 'class-transformer';
 import { EventMessage } from '../../event-message';
-import { SensorRegistered as V1 } from './1.0.0/sensor-registered.event';
-import { SensorRegistered as V2 } from './1.0.1/sensor-registered.event';
+import { SensorRegistered as V1 } from './sensor-registered-v1.event';
+import { SensorRegistered as V2 } from './sensor-registered-v2.event';
 
-export { SensorRegistered } from './1.0.1/sensor-registered.event';
+export { SensorRegistered } from './sensor-registered-v2.event';
 
 export function getSensorRegisteredEvent(eventMessage: EventMessage): V2 {
     if (!eventMessage.metadata.version || eventMessage.metadata.version === V1.version) {
