+++ src/sync/core/events/sensor/registered/sensor-registered-v2.event.ts
@@ -1,10 +1,10 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 import { Exclude, Expose } from 'class-transformer';
 
 @Exclude()
 export class SensorRegistered extends SensorEvent {
 
-  static version = '1.0.1';
+  static version = '2';
 
   @Expose()
   readonly organizationId: string;
