+++ src/sync/core/events/sensor/relocated/sensor-relocated-v1.event.ts
@@ -1,10 +1,10 @@
-import { SensorEvent } from '../../sensor.event';
+import { SensorEvent } from '../sensor.event';
 import { Exclude, Expose } from 'class-transformer';
 
 @Exclude()
 export class SensorRelocated extends SensorEvent {
 
-  static version = '1.0.0';
+  static version = '1';
 
   @Expose()
   readonly longitude: number;
