+++ src/sync/core/events/sensor/sensor-datastreamchanged.event.ts
@@ -4,7 +4,7 @@ import { Exclude, Expose } from 'class-transformer';
 @Exclude()
 export abstract class DatastreamChanged extends SensorEvent {
 
-  static version = '1.0.0';
+  static version = '1';
 
   @Expose()
   public readonly dataStreamId: string;
