+++ src/sync/core/events/sensor/updated/index.ts
@@ -1,8 +1,8 @@
 import { plainToClass } from 'class-transformer';
 import { EventMessage } from '../../event-message';
-import { SensorUpdated as V1 } from './1.0.0/sensor-updated.event';
+import { SensorUpdated as V1 } from './sensor-updated-v1.event';
 
-export { SensorUpdated } from './1.0.0/sensor-updated.event';
+export { SensorUpdated } from './sensor-updated-v1.event';
 
 export function getSensorUpdatedEvent(eventMessage: EventMessage): V1 {
     return !eventMessage.metadata.version || eventMessage.metadata.version === V1.version ? plainToClass(V1, eventMessage.data) : null;
